Refactoring, preparing to spliting the code.
